+++ src/components/molecules/index.ts
@@ -1,4 +1,5 @@
 export { default as Header } from "./Header";
-export { default as Footer } from "./Footer";
 export { default as SearchModal } from "./SearchModal";
+export { default as MobileHeader } from "./MobileHeader";
+export { default as DesktopHeader } from "./DesktopHeader";
 export { default as MobileHamburger } from "./MobileHamburger";
